autogen and apply psp patch

+++ include/pic.h
@@ -44,12 +44,12 @@ INLINE Bits PIC_TickIndexND(void) {
 	return CPU_CycleMax-CPU_CycleLeft-CPU_Cycles;
 }
 
-INLINE Bits PIC_MakeCycles(double amount) {
+INLINE Bits PIC_MakeCycles(float amount) {
 	return (Bits)(CPU_CycleMax*amount);
 }
 
-INLINE double PIC_FullIndex(void) {
-	return PIC_Ticks+(double)PIC_TickIndex();
+INLINE float PIC_FullIndex(void) {
+	return PIC_Ticks+(float)PIC_TickIndex();
 }
 
 void PIC_ActivateIRQ(Bitu irq);
